moved `<fstream>` wrapper to `ustream.hpp`
